refactoring: TermAgreement 리팩토링
- style 파일 분리하기
- PayTermInformation -> TermAgreement 로 이름 변경하기
