+++ src/pages/Payment/components/TermAgreement/index.style.ts
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const TermInformationTop = styled.div`
+  display: flex;
+  gap: 10px;
+
+  font-size: 14px;
+`;
+
+export const TermInformationBottom = styled.div`
+  display: flex;
+  justify-content: space-between;
+
+  font-size: 14px;
+`;
