+++ src/pages/Payment/components/index.ts
@@ -1,3 +1,3 @@
 export { default as PaymentAmount } from './PaymentAmount';
 export { default as CardSelection } from './CardSelection';
-export { default as PayTermInformation } from './PayTermInformation';
+export { default as TermAgreement } from './TermAgreement';
