+++ src/pages/Payment/index.style.ts
@@ -21,17 +21,3 @@ export const ButtonContainer = styled.div`
 export const ButtonLabel = styled.div`
   padding: 5px;
 `;
-
-export const TermInformationTop = styled.div`
-  display: flex;
-  gap: 10px;
-
-  font-size: 14px;
-`;
-
-export const TermInformationBottom = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  font-size: 14px;
-`;
